+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
-- Adds first version
+## [1.0.0] - 2024-12-31
+
+### Added
+
+- Numeric code generation.
+- Alpha code generation.
+- AlphaNumeric code generation.
+- Password code generation.
+- Code verification.